test: Make unwinder test work on windows (#141)
* avoid windows.h include to make armin happy, and fix build warnings on windows

* fix unwinder test on windows

* remove old cpp tests

* fix(windows): Fix TEST_VISIBLE for MSVC

* meta(vscode): Update and clean up settings

* meta(ignore): Add VisualStudio files to gitignore

* meta(vscode): Add recommended extensions

* style: Run make format

+++ tests/test_value.c
@@ -153,7 +153,7 @@ SENTRY_TEST(value_object)
     for (size_t i = 0; i < 10; i++) {
         char key[100];
         sprintf(key, "key%d", (int)i);
-        sentry_value_set_by_key(val, key, sentry_value_new_int32(i));
+        sentry_value_set_by_key(val, key, sentry_value_new_int32((int32_t)i));
     }
     for (size_t i = 0; i < 20; i++) {
         char key[100];
